total redesign - extracted code into files
