+++ src/SplashScreen.tsx
@@ -1,0 +1,7 @@
+export const SplashScreen = () => {
+  return (
+    <div className='flex items-center justify-center h-full'>
+    <p className='text-9xl font-serif'>allie.</p>
+    </div>
+  )
+};
